pyproject.toml
[project]
name = "cybertracker"
version = "0.1.0"
description = "CyberTracker REST API"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi",
    "scalar-fastapi",
    "uvicorn[standard]",
    "yarl",
    "python-multipart",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pydantic",
    "pydantic-settings",
    "bcrypt",
    "python-jose[cryptography]",
    "slowapi",
    "scikit-learn",
    "joblib",
    "numpy",
]

[dependency-groups]
dev = [
    "ruff",
    "ty",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py314"
line-length = 100
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.ty.environment]
root = ["./app"]
python-version = "3.14"

[tool.ty.src]
include = ["app"]
